adversarial_project/requirements.txt
fastapi
uvicorn
pydantic
torch
torchvision
Pillow
streamlit
requests
